Get default replacements from callback
Instead of assigning a dictionary of default replacements to a `TimeSeriesExtractorOptions` instance, a callback is now used instead.

This allows the default replacement to be dynamically supplied by the consumer instead of up-front. The use case for this is when parsing messages received via e.g. Azure Event Hub or RabbitMQ, where properties/headers from the message could be used to supply replacement values.