Add custom start position
Allow a custom start position for processing to be specified in `TimeSeriesExtractorOptions`.

The start position is specified as a JSON pointer. The property name for a sample in the document will not include this prefix i.e. if the start position is `/data` and a property is found at `/data/temperature`, the name of the property is considered to be `/temperature`.

Specifying a start position also affects the `TimestampProperty`, as it becomes relative to the start position.