Add JsonPointerMatch
`JsonPointerMatch` is intended to hold match rules when defining JSON pointers to process or skip when processing a JSON document.

It can define a `JsonPointer` to match against but it does not have to; it also allows a pattern match expression to be specified that does not have to be a syntactically valid pointer.