Fix recursion depth check in match delegates
When calling `TimeSeriesExtractor.CreateJsonPointerMatchDelegate`, the generated delegate did not correctly check the recursion depth in scenarios where the JSON pointer segment count is less than the match segment count, and the match segment count is the same as the maximum recursion depth.

This happened because the element stack on the extractor context always contains the root element in the document, which meant that the current recursion depth was being treated as one greater than it actually was.